% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Censored_data.R
\name{Censored_data}
\alias{Censored_data}
\title{Censored Data}
\usage{
Censored_data(df, res = Result4IR, resqual = ResultOp4IR,
  crit = crit)
}
\arguments{
\item{df}{dataframe to modify.}

\item{res}{column in dataframe with the result to be compared against. Defaults to Result4IR}

\item{resqual}{column in dataframe for the qualifier. Defaults to ResultOp4IR}

\item{crit}{column in dataframe with criteria value. Defaults to crit}
}
\description{
This function deals with censored data as described in the white paper "Use of Censored Data" from 5/11/2018
When the QL > criteria value, 1/2 the value of the criteria is substituted for any sample reported as censored
When the QL < Criteria value, 1/2 the QL value will be substituted
Samples greater than the Max QL, use QL value
Input is a dataframe
Adds a column named Result_cen for the value to use
Column names should be surronded by backticks ``
}
