% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Split.R
\name{Data_Split}
\alias{Data_Split}
\title{Data Split function}
\usage{
Data_Split(df, n, filepath)
}
\arguments{
\item{df}{data frame to be split}

\item{n}{row size to split dataframe into}

\item{filepath}{pathway to save file. Must end with /}
}
\description{
This function splits a data frame (df) into files of size n and saves csv to filepath
}
\examples{

}
