% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Split_AWQMS.R
\name{Data_Split_AWQMS}
\alias{Data_Split_AWQMS}
\title{Data Split function}
\usage{
Data_Split_AWQMS(df, split_on, size, filepath)
}
\arguments{
\item{df}{data frame to be split}

\item{split_on}{name of the column in dataframe to be split on}

\item{size}{row length to trigger splitting. Files will be longer than this length}

\item{filepath}{string for pathway to save file. Must end with /}
}
\description{
This function splits a data frame (df) into files of size n and saves csv to filepath
}
\examples{

}
