% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_Dup_remover.R
\name{DO_Dup_remover}
\alias{DO_Dup_remover}
\title{DO_dup_remover}
\usage{
DO_Dup_remover(df, filename = "Parameters/DO/Duplicated_DOsat_data.csv")
}
\arguments{
\item{df}{Dataframe to search for duplicates. This should be run right 
before the IR_export function of the data}

\item{filename}{location and filename for the csv of generated duplcate values}
}
\description{
This function removes duplcated DO data from further analysis. AWQMS has some
duplicate DOsat values that needs to be removed. Until this is sorted out,
DO concentration values get duplcaied when the code joins them with the 
duplcaied DOsat value. This function will identify these duplcate values
and keep only the first instance. 
Input is a dataframe and a save location for the saved duplicate file
}
