% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_check.R
\name{dup_check}
\alias{dup_check}
\title{dup_check}
\usage{
dup_check(df, chr)
}
\arguments{
\item{df}{dataframe for dup check. This should be the dataframe imported from IR database}

\item{chr}{parameter name used for output filename only}
}
\description{
This function perfroms a duplicate check based on 
MLocID, ActStartD, ActStartT, ChrName and Results
}
\examples{
 
}
