% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions_conv.R
\name{excursions_conv}
\alias{excursions_conv}
\title{excursions_conv}
\usage{
excursions_conv(n)
}
\arguments{
\item{n}{number of results}
}
\description{
This function returns the number of excursions required to list as impaired for conventional pollutants,
}
\examples{
excursions_conv()
}
