% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions_tox.R
\name{excursions_tox}
\alias{excursions_tox}
\title{excursions_tox}
\usage{
excursions_tox(n)
}
\arguments{
\item{n}{number of results}
}
\description{
This function returns the number of excursions required to list as impaired for toxic pollutants,
}
\examples{
excursions_tox()
}
