% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean.R
\name{geo_mean}
\alias{geo_mean}
\title{geo_mean}
\usage{
geo_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector to calculate geometric mean}
}
\description{
This function calculates a geometric mean for a vector
borrowed from https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in
Input is name of vector
}
