% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_export.R
\name{IR_export}
\alias{IR_export}
\title{IR categories export function function}
\usage{
IR_export(df, location, parameter)
}
\arguments{
\item{df}{data frame to be split}

\item{location}{pathway to put exported files}

\item{parameter}{parameter of export file such as "DO_yearround_continuous"}
}
\description{
This function splits a data frame by column named OWRD_Basin
}
\examples{

}
