% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_Validation.R
\name{IR_Validation}
\alias{IR_Validation}
\title{IR_Validation}
\usage{
IR_Validation(Results_import, anom_crit, parameter)
}
\arguments{
\item{Results_import}{dataframe for the imported results}

\item{anom_crit}{dataframe for the ambient program criteria}

\item{parameter}{parameter classification used for naming invalid data file}
}
\description{
This function performs the IR data validation process. Results are compared to oregon ambient
data for 1980 - 2017. Data outside the 99th percentile are labeled invalid. A table of Invalid data
is written to Parameters/Invalid_data. Assessment Units with no invalid data are returned as dataframe.
Input is the imported results dataframe, the anom_crit dataframe describing the oregon ambient data,
and the parameter classification e.g. Bacteria, pH. etc.
}
