% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR_Validation_Import.R
\name{IR_Validation_Import}
\alias{IR_Validation_Import}
\title{IR_Validation_Import}
\usage{
IR_Validation_Import(file, df)
}
\arguments{
\item{file}{File that contains reviewd data validation table}

\item{df}{Dataframe that results from IR_Validation function. For the IR, this will be the results of the datapull script.}
}
\description{
This function rejoins the data that was removed from the process during the data validation step. 
IR_Validation creates a table that must be reviewed. If data labled as invalid is actually valid,
change the Conclusion field to read "Valid"
}
